algorithm implementation

+++ src/placer/legalizer/LGTetris/tetrisspace.h
@@ -30,7 +30,7 @@ private:
 public:
   TetrisSpace(int left, int right, int bottom, int top, int rowHeight, int minWidth);
   void FindCommonSegments(int startRowNum, int endRowNum, FreeSegmentList &commonSegments);
-  bool IsSpaceAvail(double x_loc, double y_loc, int width, int height);
+  bool IsSpaceAvail(int x_loc, int y_loc, int width, int height);
   bool FindBlockLoc(double currentX, double currentY, int blockWidth, int blockHeight, Loc2D result_loc);
   void Show();
 };
